requirements.txt
phonenumbers>=8.13.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
rich>=13.0.0
python-dotenv>=1.0.0
